fix(linked-list): double check psuedoCode mapping and fix it
